+++ src/Designer/frontend/admin/routes/RoutePathsParams.ts
@@ -1,0 +1,6 @@
+export type RoutePathsParams = {
+  owner: string;
+  environment: string;
+  app: string;
+  instanceId: string;
+};