+++ src/Designer/frontend/settings/features/orgs/components/Menu/Menu.tsx
@@ -4,23 +4,29 @@ import { useNavigate, useLocation } from 'react-router-dom';
 import { RobotSmileIcon, BellIcon } from '@studio/icons';
 import { useTranslation } from 'react-i18next';
 import { RoutePaths } from '../../routes/RoutePaths';
+import { FeatureFlag, useFeatureFlag } from '@studio/feature-flags';
 
 export function Menu(): ReactElement {
   const { t } = useTranslation();
   const navigate = useNavigate();
   const { pathname } = useLocation();
   const selectedTabId = pathname.split('/').at(-1);
+  const adminEnabled = useFeatureFlag(FeatureFlag.Admin);
   const menuTabs = [
     {
       tabId: RoutePaths.BotAccounts,
       tabName: t('settings.orgs.bot_accounts.menu.bot_accounts'),
       icon: <RobotSmileIcon />,
     },
-    {
-      tabId: RoutePaths.ContactPoints,
-      tabName: t('settings.orgs.contact_points.menu.contact_points'),
-      icon: <BellIcon />,
-    },
+    ...(adminEnabled
+      ? [
+          {
+            tabId: RoutePaths.ContactPoints,
+            tabName: t('settings.orgs.contact_points.menu.contact_points'),
+            icon: <BellIcon />,
+          },
+        ]
+      : []),
   ];
   return (
     <StudioContentMenu